Merged PR 1382: Fake WRF loop upper bounds
Fake WRF loop upper bounds by introducing explicit variables, which also ease the control the magnitude of the unrolled graph.

However, due to the complicity of  the WRF model itself, even to reduce the faked UBs to 1, 1, 1, there will still be ~4200 nodes. (in contrast, the original UBs result in ~23000 nodes).