Merged PR 1395: Add select_scatter runtime function
PyTorch v1.11 add 'torch.select_scatter' function, which has the most concise and accurate semantics.
But the torch-to-ONNX component in the same version doesn't handle this function well.
Two other alternatives: in-place `x[i,:] = sub`, which is in-compatible with current dataflow; 'y = x.clone(); y[i,:] = sub` causes crash.
So instead of these, for codegen to use, add a runtime function based on `torch.masked_scatter` which is well supported in ONNX opset=11.