Merged PR 1902: reverse auto scale order of multiref and pyop
reverse auto scale order of multiref and pyop

This was originally a bug when we firstly scale multiref and then pyop: if the multiref depends on pyop, the pyop is not scaled so the multiref will be kept as same.

Since pyop now will always be replicated to all the devices, we can put pyop first because it depends nothing. Then the multiref can be correctly scaled after the scale of pyop.