Merged PR 2043: parallel module: remove reducer_op option
reducer_op option is not very useful and can lead to inconsistence when there is intra scale unit reduce.

For example,
If a param is replicated due to op's partition, then the param will be added to reducer and its reduce behavior depends on reducer_op option.

But if the param is used by multiple ops, and some ops are partitioned, and other ops are replicated,  In that case, the param will be reduced by identity-allreduce, and the reducer_op is always `sum` (regardless the value of reducer_op option)

So let's remove this option.