Merged PR 2102: Fix persistent buffer
The original logic is to register all non-nn.Parameter tensors as persistent buffers, and then these buffers will be saved together in the checkpoint. On the one hand, this will lead to useless checkpoint storage, and some buffers will be stored in fp32 because they are initialized to fp32. When restoring in fp16 format, these tensors will be loaded into fp16 format. In the process of fp32_dump, load ckpt, and to_fp16, checkpoint comparison will be inconsistent due to accuracy issues.

Now, ordinary tensors will be kept registered as non-persistent buffers, and the original buffers will remain persistent and consistent with the original.