Merged PR 2116: fix dimops align infer
in dimops anno infer, we should support the following case:

(a b), a -> a b

in previous, if a tensor has combined identifiers like (a b), then we will directly tell user we cannot infer the unknown value of a b and raise error.

but in fact, we can get a and b value from other tensor identifier, this pr support this case.