Merged PR 2223: refine pytree
in this pr, directly copy torch 2.3 pytree as `_pytree.py`, it is more powerful than the previous version, and implemented most of the functions we needed, the only reason to copy the file instead of import is we also support version 2.0 <= torch < 2.3. there is no need to review `_pytree.py`.

the additional support for pytree is in `pytree_utils.py`

parity check passed.