Merged PR 2309: [Bugfix] tracer: handle importlib carefully
Currently we only wrap function calls via `ProxyCallTransformer`, but some functions can be triggered by getattr (e.g. torch._dynamo), in which it is impossible to 'unpatch' the calls triggered inside `getattr` (Make it leaf function doesn't work because we don't have opportunity to unpatch it).

This PR is trying to fix a special case when `importlib.import_module` is called inside `getattr`.
Two reasons:
1. import_module will run the code of modules recursively (the imported module can import other modules), and is easy to have problem.
2. `torch._dynamo` is used in some pytorch modules, and we want to support it.

Two ways to fix this:
 1. Handle `import_module` in a special way.
 2. Refine the `ProxyCallTransformer` to patch getattr as well.

The second way is more general, but it's more complex and may introduce potential bugs.

In this PR, we choose the first way as a quick fix.