Merged PR 2432: [Bugfix] return cloned tensor when a parameter is multiref'ed in zero3
The original multiref will return the parameter itself, which doesn't work when zero3 is enabled.
Because after multiref returns, the parameter will be evicted from memory, so will the return values.

To fix this problem, we add a new argument(clone_level) to multiref, and return a clone for parameters when zero3 is enable.